requirements.txt
# requirements.txt - 개선된 라이브러리 목록

#===================================================================
# COCONUT: Continual & Competition Network with Updatable Templates

# 프로젝트 실행에 필요한 라이브러리 목록입니다.

# INSTALLATION:
# pip install -r requirements.txt

# 🔥 IMPORTANT: Choose either faiss-gpu OR faiss-cpu (not both)
#===================================================================

# --- 1. 핵심 딥러닝 라이브러리 ---
torch==1.10.0
torchvision==0.11.1
setuptools==58.2.0

# --- 2. 🔥 COCONUT 핵심 라이브러리 ---
# GPU 사용 시 (권장)
faiss-gpu>=1.7.0
# CPU 사용 시 (위 줄을 주석처리하고 아래 줄 사용)
# faiss-cpu>=1.7.0

# YAML 설정 파일 파싱
pyyaml>=5.4.0

# --- 3. 과학 계산 및 데이터 분석 ---
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# --- 4. 이미지 처리 ---
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# --- 5. 유틸리티 및 시각화 ---
tqdm>=4.62.0
matplotlib>=3.4.0

# --- 6. W2ML 수학적 검증용 (선택사항) ---
# sympy>=1.8.0  # 수식 검증용
# jupyter>=1.0.0  # 실험 노트북용

# ===================================================================
# INSTALLATION GUIDE:
#
# 1. GPU 환경 (권장):
#    pip install torch==1.10.0 torchvision==0.11.1 -f https://download.pytorch.org/whl/cu111/torch_stable.html
#    pip install faiss-gpu
#    pip install -r requirements.txt
#
# 2. CPU 환경:
#    pip install torch==1.10.0 torchvision==0.11.1 -f https://download.pytorch.org/whl/cpu/torch_stable.html
#    pip install faiss-cpu
#    pip install -r requirements.txt
#
# 3. 검증:
#    python -c "import torch, faiss; print('✅ Installation successful')"
# ===================================================================